Civilization: Remove vSync & use getDeltaTime()

+++ core/src/main/java/me/rhin/openciv/ui/screen/AbstractScreen.java
@@ -231,7 +231,7 @@ public abstract class AbstractScreen implements Screen, InputProcessor {
 		this.camY = camY;
 	}
 
-	public void translateCamera(int x, int y, int z) {
+	public void translateCamera(float x, float y, float z) {
 		camX += x;
 		camY += y;
 	}
